requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
alabaster==1.0.0
albucore==0.0.20
albumentations==1.4.21
alembic==1.15.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
autowoe==1.3.2
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.9.0
boto3==1.35.49
botocore==1.35.49
cachetools==5.5.0
catboost==1.2.7
category-encoders==2.6.4
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
cmaes==0.11.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
colour==0.1.5
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
dataclasses-json==0.6.7
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.2
docutils==0.21.2
dotenv==0.9.9
dtaidistance==2.3.12
dtreeviz==2.2.2
duckduckgo_search==8.1.1
easyocr==1.7.1
efficientnet_pytorch==0.7.1
et-xmlfile==1.1.0
eval_type_backport==0.2.0
executing==2.1.0
fastdtw==0.3.4
fastjsonschema==2.20.0
filelock==3.15.4
fire==0.6.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gensim==4.3.3
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-vision==3.8.1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grandalf==0.8
graphviz==0.20.3
greenlet==3.2.0
grpcio==1.71.0
grpcio-status==1.67.0
h11==0.14.0
h5py==3.13.0
holidays==0.61
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
image==1.5.33
imageio==2.35.1
imagesize==1.4.1
imgaug==0.4.0
iniconfig==2.0.0
intel-openmp==2021.4.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
JPype1==1.5.0
json2html==1.3.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.9.2
kiwisolver==1.4.5
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-experimental==0.3.4
langchain-mcp-adapters==0.1.7
langchain-openai==0.3.24
langchain-text-splitters==0.3.8
langgraph==0.5.1
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.70
langsmith==0.3.45
lazy_loader==0.4
libclang==18.1.1
lightautoml==0.2.13
lightgbm==2.3.1
llvmlite==0.43.0
lmdb==1.5.1
log_calls==0.3.2
lxml==5.3.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mcp==1.9.4
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
ml_dtypes==0.5.1
mljar-scikit-plot==0.3.12
mljar-supervised==1.1.14
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==2.1.3
omegaconf==2.3.0
onnx==1.16.2
onnx-graphsurgeon==0.5.2
onnx-tf==1.6.0
onnx2tf==1.22.3
onnxruntime==1.19.2
onnxslim==0.1.34
openai==1.88.0
openai-whisper==20231117
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
optuna==4.1.0
optuna-integration==4.1.0
orjson==3.10.7
ormsgpack==1.10.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
paddleocr==2.8.1
paddlepaddle==2.6.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pdf2image==1.17.0
piexif==1.1.3
pillow==11.2.1
pillow_heif==0.22.0
platformdirs==4.2.2
plotly==5.24.1
pluggy==1.5.0
poetry-core==1.9.1
primp==0.15.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.29.4
psutil==6.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.9.1
pydantic_core==2.27.1
pygame==2.6.1
Pygments==2.19.1
PyMuPDF==1.24.10
pymupdf4llm==0.0.16
PyMuPDFb==1.24.10
pyparsing==3.1.4
pypdfium2==4.30.0
pyreadline3==3.5.2
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.3
python-bidi==0.6.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.1
PyWavelets==1.7.0
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.15.2
seaborn==0.13.2
selenium==4.25.0
Send2Trash==1.8.3
setuptools==78.1.0
shap==0.46.0
shapely==2.0.6
simsimd==6.0.7
six==1.17.0
slicer==0.0.8
smart-open==7.0.5
sng4onnx==1.0.4
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.40
sqlparse==0.5.1
sse-starlette==2.3.4
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
StrEnum==0.4.15
stringzilla==3.10.10
sympy==1.13.1
tabula-py==2.9.3
tabulate==0.9.0
tavily-python==0.4.0
tbb==2021.13.1
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-intel==2.17.0
termcolor==3.0.1
terminado==0.18.1
tf_keras==2.17.0
tflite-support==0.1.0a1
threadpoolctl==3.5.0
tifffile==2024.8.30
tiktoken==0.7.0
timm==1.0.11
tinycss2==1.3.0
tokenizers==0.20.3
torch==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.3
trio==0.26.2
trio-websocket==0.11.1
tslearn==0.6.3
types-python-dateutil==2.9.0.20240821
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.1
ultralytics==8.2.86
ultralytics-thop==2.0.6
uri-template==1.3.0
urllib3==2.4.0
uv==0.7.3
uvicorn==0.34.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.1.2
xxhash==3.5.0
yarl==1.9.7
yomitoku==0.5.1
zstandard==0.23.0
